requirements.txt
accelerate==1.12.0
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
etelemetry==0.3.1
faiss-cpu==1.13.2
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-vision==3.12.1
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.11
langchain-openai==1.1.9
langchain-text-splitters==1.1.0
langdetect==1.0.9
langsmith==0.7.1
llvmlite==0.46.0
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.24.1
openai==2.20.0
opencv-python==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pi_heif==1.2.0
pikepdf==10.3.0
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
prov==2.1.1
psutil==7.2.2
puremagic==1.30
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==5.4.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
RapidFuzz==3.14.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==82.0.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traits==7.1.0
transformers==5.1.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.31
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
webencodings==0.5.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
